+++ FeatureExtractionNew/FeatureExtractionNew.ino/Cmplx.h
@@ -4,15 +4,15 @@
 namespace co {
 
 	typedef struct my_complex {
-		double real;
-		double imag;
+		float real;
+		float imag;
 
 		my_complex() {
 			real = 0;
 			imag = 0;
 		}
 
-		my_complex(double r, double i) {
+		my_complex(float r, float i) {
 			real = r;
 			imag = i;
 		}
@@ -21,7 +21,7 @@ namespace co {
 
 	my_complex add(cd ,cd);
 	my_complex sub(cd, cd);
-	my_complex mul(cd, double);
+	my_complex mul(cd, float);
 	my_complex mul(cd, cd);
 	my_complex e(cd);
 	std::string toString(cd);
