+++ FeatureExtractionNew/FeatureExtractionNew.ino/ExtractionHandler.h
@@ -14,53 +14,53 @@ namespace eh {
 
 	public:
 		
-		double handle_mean(std::string, std::vector<double>&);
-		double handle_mean_abs_dev(std::string, std::vector<double>&);
-		double handle_mean_geometric_abs(std::string, std::vector<double>&);
-		double handle_median(std::string, std::vector<double>&);
-		double handle_median_abs_changes(std::string, std::vector<double>&);
-		double handle_median_changes(std::string, std::vector<double>&);
-		double handle_median_abs_dev(std::string, std::vector<double>&);
-		double handle_std_dev(std::string, std::vector<double>&);
-		double handle_avg_dev(std::string, std::vector<double>&);
-		double handle_var(std::string, std::vector<double>&);
-		double handle_abs_energy(std::string, std::vector<double>&);
-		double handle_kurtosis(std::string, std::vector<double>&);
-		double handle_skewness(std::string, std::vector<double>&);
-		double handle_zero_cross(std::string, std::vector<double>&);
+		float handle_mean(std::string, std::vector<float>&);
+		float handle_mean_abs_dev(std::string, std::vector<float>&);
+		float handle_mean_geometric_abs(std::string, std::vector<float>&);
+		float handle_median(std::string, std::vector<float>&);
+		float handle_median_abs_changes(std::string, std::vector<float>&);
+		float handle_median_changes(std::string, std::vector<float>&);
+		float handle_median_abs_dev(std::string, std::vector<float>&);
+		float handle_std_dev(std::string, std::vector<float>&);
+		float handle_avg_dev(std::string, std::vector<float>&);
+		float handle_var(std::string, std::vector<float>&);
+		float handle_abs_energy(std::string, std::vector<float>&);
+		float handle_kurtosis(std::string, std::vector<float>&);
+		float handle_skewness(std::string, std::vector<float>&);
+		float handle_zero_cross(std::string, std::vector<float>&);
 		
-		double handle_max(std::string, std::vector<double>&);
-		double handle_abs_max(std::string, std::vector<double>&);
-		double handle_min(std::string, std::vector<double>&);
-		double handle_last_location_of_max(std::string, std::vector<double>&);
-		double handle_last_location_of_min(std::string, std::vector<double>&);
-		double handle_first_location_of_max(std::string, std::vector<double>&);
-		double handle_first_location_of_min(std::string, std::vector<double>&);
-		double handle_mean_n_abs_max(std::string, std::vector<double>&, double);
+		float handle_max(std::string, std::vector<float>&);
+		float handle_abs_max(std::string, std::vector<float>&);
+		float handle_min(std::string, std::vector<float>&);
+		float handle_last_location_of_max(std::string, std::vector<float>&);
+		float handle_last_location_of_min(std::string, std::vector<float>&);
+		float handle_first_location_of_max(std::string, std::vector<float>&);
+		float handle_first_location_of_min(std::string, std::vector<float>&);
+		float handle_mean_n_abs_max(std::string, std::vector<float>&, float);
 
-		double handle_mean_abs_changes(std::string, std::vector<double>&);
-		double handle_mean_changes(std::string, std::vector<double>&);
-		double handle_abs_sum_of_changes(std::string, std::vector<double>&);
-		double handle_change_quantile(std::string, std::vector<double>&, double, double, double);
+		float handle_mean_abs_changes(std::string, std::vector<float>&);
+		float handle_mean_changes(std::string, std::vector<float>&);
+		float handle_abs_sum_of_changes(std::string, std::vector<float>&);
+		float handle_change_quantile(std::string, std::vector<float>&, float, float, float);
 
-		double handle_sum(std::string, std::vector<double>&);
-		double handle_range_count(std::string, std::vector<double>&, double, double);
-		double handle_non_zero_count(std::string, std::vector<double>&);
-		double handle_count_above(std::string, std::vector<double>&, double);
-		double handle_count_above_mean(std::string, std::vector<double>&);
-		double handle_count_below(std::string, std::vector<double>&, double);
-		double handle_count_below_mean(std::string, std::vector<double>&);
-		double handle_root_mean_square(std::string, std::vector<double>&);
-		double handle_quantile(std::string, std::vector<double>&, double);
-		double handle_interquartile_range(std::string, std::vector<double>&);
-		double handle_negative_turnings(std::string, std::vector<double>&);
-		double handle_positive_turnings(std::string, std::vector<double>&);
+		float handle_sum(std::string, std::vector<float>&);
+		float handle_range_count(std::string, std::vector<float>&, float, float);
+		float handle_non_zero_count(std::string, std::vector<float>&);
+		float handle_count_above(std::string, std::vector<float>&, float);
+		float handle_count_above_mean(std::string, std::vector<float>&);
+		float handle_count_below(std::string, std::vector<float>&, float);
+		float handle_count_below_mean(std::string, std::vector<float>&);
+		float handle_root_mean_square(std::string, std::vector<float>&);
+		float handle_quantile(std::string, std::vector<float>&, float);
+		float handle_interquartile_range(std::string, std::vector<float>&);
+		float handle_negative_turnings(std::string, std::vector<float>&);
+		float handle_positive_turnings(std::string, std::vector<float>&);
 
-		double handle_autocorrelation(std::string, std::vector<double>&, double);
-		std::vector<double> handle_mfcc(std::string, std::vector<double>&, double, double, double);
-		std::vector<co::cd> handle_fft(std::string, std::vector<double>&);
-		std::vector<double> handle_lpc(std::string, std::vector<double>&, double, double);
-		std::vector<double> handle_lpcc(std::string, std::vector<double>&, double, double, double);
+		float handle_autocorrelation(std::string, std::vector<float>&, float);
+		std::vector<float> handle_mfcc(std::string, std::vector<float>&, float, float, float);
+		std::vector<co::cd> handle_fft(std::string, std::vector<float>&);
+		std::vector<float> handle_lpc(std::string, std::vector<float>&, float, float);
+		std::vector<float> handle_lpcc(std::string, std::vector<float>&, float, float, float);
 
 	};
